+++ backend/index.js
@@ -11,25 +11,16 @@ dbUrl = process.env.DBURI || "mongodb://localhost:27017/test";
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
 
-
-app.get('/', function(req, res) {
-    res.send('Hello world');
-});
+require("./routes/user")(app);
+require("./routes/parkingSpot")(app);
 
 db.connect(dbUrl).then(() => {
     console.log("Connected to the database!");
+    app.listen(PORT, function() {
+      console.log('Server is running on port:', PORT);
+    });
   })
   .catch(err => {
     console.log("Cannot connect to the database!", err);
     process.exit();
-  });
-
-
-require("./routes/user.route")(app);
-
-
-
-//Server listening
-app.listen(PORT, function() {
-    console.log('Server is running on port:', PORT);
 });
