+++ backend/routes/parkingSpot.js
@@ -1,6 +1,6 @@
 module.exports = app => {
       
-    const parkingSpots = require("../controllers/parkingSpot.controller.js");
+    const parkingSpots = require("../controllers/parkingSpot.js");
     var router = require("express").Router();
     router.post("/", parkingSpots.create);
