+++ backend/routes/user.js
@@ -1,9 +1,9 @@
 module.exports = app => {
       
-    const users = require("../controllers/user.controller.js");
+    const users = require("../controllers/user.js");
     var router = require("express").Router();
     router.post("/signup", users.create);
 
   
     app.use("/", router);
-  };
+};
